+++ src/main/java/com/KDJStudios/XMPPJabberClient/crypto/axolotl/AxolotlService.java
@@ -58,7 +58,7 @@ import com.KDJStudios.XMPPJabberClient.xmpp.stanzas.IqPacket;
 
 public class AxolotlService implements OnAdvancedStreamFeaturesLoaded {
 
-	public static final String PEP_PREFIX = "eu.siacs.conversations.axolotl";
+	public static final String PEP_PREFIX = "com.KDJStudios.XMPPJabberClient.axolotl";
 	public static final String PEP_DEVICE_LIST = PEP_PREFIX + ".devicelist";
 	public static final String PEP_DEVICE_LIST_NOTIFY = PEP_DEVICE_LIST + "+notify";
 	public static final String PEP_BUNDLES = PEP_PREFIX + ".bundles";