+++ src/playstore/java/eu/siacs/conversations/services/MaintenanceReceiver.java
@@ -15,7 +15,7 @@ public class MaintenanceReceiver extends BroadcastReceiver {
 	@Override
 	public void onReceive(Context context, Intent intent) {
 		Log.d(Config.LOGTAG,"received intent in maintenance receiver");
-		if ("eu.siacs.conversations.RENEW_INSTANCE_ID".equals(intent.getAction())) {
+		if ("com.KDJStudios.XMPPJabberClient.RENEW_INSTANCE_ID".equals(intent.getAction())) {
 			renewInstanceToken(context);
 
 		}