+++ CHANGELOG.md
@@ -427,7 +427,7 @@
 * bug fixes
 
 ### Version 1.2.0
-* Send current location. (requires [plugin](https://play.google.com/store/apps/details?id=eu.siacs.conversations.sharelocation))
+* Send current location. (requires [plugin](https://play.google.com/store/apps/details?id=com.KDJStudios.XMPPJabberClient.sharelocation))
 * Invite multiple contacts at once
 * performance improvements
 * bug fixes